requirements.txt
bcrypt
bleach
email_validator
Flask
Flask-Executor
Flask-mailman
Flask-Migrate
flask-minify
Flask-Reuploaded
Flask-Security-Too
Flask-Session
Flask-SQLAlchemy
flask-uploads
Flask-WTF
loguru
PyMySql
python-dotenv
requests
git+https://github.com/jokerinteractive/flask-resize@master
